refactoring of JAXRandomKey and more to support private API
