+++ src/pyggdrasil/tree_inference/_mcmc.py
@@ -15,9 +15,10 @@ import dataclasses
 import logging
 
 
+from pyggdrasil.interface import JAXRandomKey
+
 from pyggdrasil.tree_inference import (
     Tree,
-    JAXRandomKey,
     get_descendants,
 )
